refactor(storage): split storage implementation into multiple files
