+++ storage/gcloud/aio/storage/__init__.py
@@ -1,10 +1,10 @@
 from pkg_resources import get_distribution
 __version__ = get_distribution('gcloud-aio-storage').version
 
-from gcloud.aio.storage.storage import Blob
-from gcloud.aio.storage.storage import Bucket
-from gcloud.aio.storage.storage import make_download
+from gcloud.aio.storage.blob import Blob
+from gcloud.aio.storage.bucket import Bucket
 from gcloud.aio.storage.storage import Storage
+from gcloud.aio.storage.utils import make_download
 
 
-__all__ = ['__version__', 'Blob', 'Bucket', 'make_download', 'Storage']
+__all__ = ['__version__', 'Blob', 'Bucket', 'Storage', 'make_download']
